# loms/api/openapiv2/loms.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Loms Service",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "OrderService"
    },
    {
      "name": "StockService"
    }
  ],
  "host": "localhost:8084",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/order": {
      "post": {
        "operationId": "OrderService_OrderCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/OrderCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderCreateRequest"
            }
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
    "/order/cancel": {
      "post": {
        "operationId": "OrderService_OrderCancel",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/OrderCancelResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderCancelRequest"
            }
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
    "/order/pay": {
      "post": {
        "operationId": "OrderService_OrderPay",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/OrderPayResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderPayRequest"
            }
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
    "/order/{orderId}": {
      "get": {
        "operationId": "OrderService_OrderInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/OrderInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
    "/stock/{skuId}": {
      "get": {
        "operationId": "StockService_StockInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/StockInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "skuId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "StockService"
        ]
      }
    }
  },
  "definitions": {
    "ItemInfo": {
      "type": "object",
      "properties": {
        "skuId": {
          "type": "string",
          "format": "int64"
        },
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "OrderCancelRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "OrderCancelResponse": {
      "type": "object"
    },
    "OrderCreateRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ItemInfo"
          }
        }
      }
    },
    "OrderCreateResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "OrderInfoResponse": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ItemInfo"
          }
        }
      }
    },
    "OrderPayRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "OrderPayResponse": {
      "type": "object"
    },
    "StockInfoResponse": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}