# loms/pkg/api/stocks/v1/stocks.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.28.1
// 	protoc        v6.31.1
// source: stocks/v1/stocks.proto

package stocks

import (
	_ "github.com/envoyproxy/protoc-gen-validate/validate"
	_ "github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-openapiv2/options"
	_ "google.golang.org/genproto/googleapis/api/annotations"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type StockInfoRequest struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	SkuId int64 `protobuf:"varint,1,opt,name=sku_id,json=skuId,proto3" json:"sku_id,omitempty"`
}

func (x *StockInfoRequest) Reset() {
	*x = StockInfoRequest{}
	if protoimpl.UnsafeEnabled {
		mi := &file_stocks_v1_stocks_proto_msgTypes[0]
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		ms.StoreMessageInfo(mi)
	}
}

func (x *StockInfoRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*StockInfoRequest) ProtoMessage() {}

func (x *StockInfoRequest) ProtoReflect() protoreflect.Message {
	mi := &file_stocks_v1_stocks_proto_msgTypes[0]
	if protoimpl.UnsafeEnabled && x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use StockInfoRequest.ProtoReflect.Descriptor instead.
func (*StockInfoRequest) Descriptor() ([]byte, []int) {
	return file_stocks_v1_stocks_proto_rawDescGZIP(), []int{0}
}

func (x *StockInfoRequest) GetSkuId() int64 {
	if x != nil {
		return x.SkuId
	}
	return 0
}

type StockInfoResponse struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	Count uint32 `protobuf:"varint,1,opt,name=count,proto3" json:"count,omitempty"`
}

func (x *StockInfoResponse) Reset() {
	*x = StockInfoResponse{}
	if protoimpl.UnsafeEnabled {
		mi := &file_stocks_v1_stocks_proto_msgTypes[1]
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		ms.StoreMessageInfo(mi)
	}
}

func (x *StockInfoResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*StockInfoResponse) ProtoMessage() {}

func (x *StockInfoResponse) ProtoReflect() protoreflect.Message {
	mi := &file_stocks_v1_stocks_proto_msgTypes[1]
	if protoimpl.UnsafeEnabled && x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use StockInfoResponse.ProtoReflect.Descriptor instead.
func (*StockInfoResponse) Descriptor() ([]byte, []int) {
	return file_stocks_v1_stocks_proto_rawDescGZIP(), []int{1}
}

func (x *StockInfoResponse) GetCount() uint32 {
	if x != nil {
		return x.Count
	}
	return 0
}

var File_stocks_v1_stocks_proto protoreflect.FileDescriptor

var file_stocks_v1_stocks_proto_rawDesc = []byte{
	0x0a, 0x16, 0x73, 0x74, 0x6f, 0x63, 0x6b, 0x73, 0x2f, 0x76, 0x31, 0x2f, 0x73, 0x74, 0x6f, 0x63,
	0x6b, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x17, 0x76, 0x61, 0x6c, 0x69, 0x64, 0x61,
	0x74, 0x65, 0x2f, 0x76, 0x61, 0x6c, 0x69, 0x64, 0x61, 0x74, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
	0x6f, 0x1a, 0x1c, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x61, 0x6e,
	0x6e, 0x6f, 0x74, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a,
	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x2d, 0x67, 0x65, 0x6e, 0x2d, 0x6f, 0x70, 0x65, 0x6e,
	0x61, 0x70, 0x69, 0x76, 0x32, 0x2f, 0x6f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x2f, 0x61, 0x6e,
	0x6e, 0x6f, 0x74, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22,
	0x32, 0x0a, 0x10, 0x53, 0x74, 0x6f, 0x63, 0x6b, 0x49, 0x6e, 0x66, 0x6f, 0x52, 0x65, 0x71, 0x75,
	0x65, 0x73, 0x74, 0x12, 0x1e, 0x0a, 0x06, 0x73, 0x6b, 0x75, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20,
	0x01, 0x28, 0x03, 0x42, 0x07, 0xfa, 0x42, 0x04, 0x22, 0x02, 0x20, 0x00, 0x52, 0x05, 0x73, 0x6b,
	0x75, 0x49, 0x64, 0x22, 0x29, 0x0a, 0x11, 0x53, 0x74, 0x6f, 0x63, 0x6b, 0x49, 0x6e, 0x66, 0x6f,
	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x63, 0x6f, 0x75, 0x6e,
	0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x05, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x32, 0x5b,
	0x0a, 0x0c, 0x53, 0x74, 0x6f, 0x63, 0x6b, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x4b,
	0x0a, 0x09, 0x53, 0x74, 0x6f, 0x63, 0x6b, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x11, 0x2e, 0x53, 0x74,
	0x6f, 0x63, 0x6b, 0x49, 0x6e, 0x66, 0x6f, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x12,
	0x2e, 0x53, 0x74, 0x6f, 0x63, 0x6b, 0x49, 0x6e, 0x66, 0x6f, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
	0x73, 0x65, 0x22, 0x17, 0x82, 0xd3, 0xe4, 0x93, 0x02, 0x11, 0x12, 0x0f, 0x2f, 0x73, 0x74, 0x6f,
	0x63, 0x6b, 0x2f, 0x7b, 0x73, 0x6b, 0x75, 0x5f, 0x69, 0x64, 0x7d, 0x42, 0x28, 0x5a, 0x26, 0x72,
	0x6f, 0x75, 0x74, 0x65, 0x32, 0x35, 0x36, 0x2f, 0x6c, 0x6f, 0x6d, 0x73, 0x2f, 0x70, 0x6b, 0x67,
	0x2f, 0x61, 0x70, 0x69, 0x2f, 0x73, 0x74, 0x6f, 0x63, 0x6b, 0x73, 0x2f, 0x76, 0x31, 0x3b, 0x73,
	0x74, 0x6f, 0x63, 0x6b, 0x73, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
}

var (
	file_stocks_v1_stocks_proto_rawDescOnce sync.Once
	file_stocks_v1_stocks_proto_rawDescData = file_stocks_v1_stocks_proto_rawDesc
)

func file_stocks_v1_stocks_proto_rawDescGZIP() []byte {
	file_stocks_v1_stocks_proto_rawDescOnce.Do(func() {
		file_stocks_v1_stocks_proto_rawDescData = protoimpl.X.CompressGZIP(file_stocks_v1_stocks_proto_rawDescData)
	})
	return file_stocks_v1_stocks_proto_rawDescData
}

var file_stocks_v1_stocks_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_stocks_v1_stocks_proto_goTypes = []interface{}{
	(*StockInfoRequest)(nil),  // 0: StockInfoRequest
	(*StockInfoResponse)(nil), // 1: StockInfoResponse
}
var file_stocks_v1_stocks_proto_depIdxs = []int32{
	0, // 0: StockService.StockInfo:input_type -> StockInfoRequest
	1, // 1: StockService.StockInfo:output_type -> StockInfoResponse
	1, // [1:2] is the sub-list for method output_type
	0, // [0:1] is the sub-list for method input_type
	0, // [0:0] is the sub-list for extension type_name
	0, // [0:0] is the sub-list for extension extendee
	0, // [0:0] is the sub-list for field type_name
}

func init() { file_stocks_v1_stocks_proto_init() }
func file_stocks_v1_stocks_proto_init() {
	if File_stocks_v1_stocks_proto != nil {
		return
	}
	if !protoimpl.UnsafeEnabled {
		file_stocks_v1_stocks_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
			switch v := v.(*StockInfoRequest); i {
			case 0:
				return &v.state
			case 1:
				return &v.sizeCache
			case 2:
				return &v.unknownFields
			default:
				return nil
			}
		}
		file_stocks_v1_stocks_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
			switch v := v.(*StockInfoResponse); i {
			case 0:
				return &v.state
			case 1:
				return &v.sizeCache
			case 2:
				return &v.unknownFields
			default:
				return nil
			}
		}
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: file_stocks_v1_stocks_proto_rawDesc,
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   1,
		},
		GoTypes:           file_stocks_v1_stocks_proto_goTypes,
		DependencyIndexes: file_stocks_v1_stocks_proto_depIdxs,
		MessageInfos:      file_stocks_v1_stocks_proto_msgTypes,
	}.Build()
	File_stocks_v1_stocks_proto = out.File
	file_stocks_v1_stocks_proto_rawDesc = nil
	file_stocks_v1_stocks_proto_goTypes = nil
	file_stocks_v1_stocks_proto_depIdxs = nil
}